# services/api/docs/swagger.yaml
definitions:
...
  lib.Response:
    properties:
      message:
        description: response message
        type: string
      status:
        description: http status
        type: integer
    type: object
...
  model.Book:
    properties:
      count:
        type: integer
      default_language:
        type: string
      hadith:
        items:
          $ref: '#/definitions/model.Hadith'
        type: array
      id:
        type: integer
      media:
        items:
          $ref: '#/definitions/model.BookAsset'
        type: array
      slug:
        type: string
      theme:
        items:
          $ref: '#/definitions/model.Theme'
        type: array
      translation:
        $ref: '#/definitions/model.Translation'
      translation_id:
        type: integer
    type: object
  model.BookAsset:
    properties:
      book:
        $ref: '#/definitions/model.Book'
      book_id:
        type: integer
      id:
        type: integer
      multimedia:
        $ref: '#/definitions/model.Multimedia'
      multimedia_id:
        type: integer
    type: object
  model.Chapter:
    properties:
      defaultLanguage:
        type: string
      id:
        type: integer
      media:
        items:
          $ref: '#/definitions/model.ChapterAsset'
        type: array
      theme:
        $ref: '#/definitions/model.Theme'
      theme_id:
        type: integer
      translation:
        $ref: '#/definitions/model.Translation'
      translation_id:
        type: integer
    type: object
  model.ChapterAsset:
    properties:
      chapter_id:
        type: integer
      id:
        type: integer
      multimedia:
        $ref: '#/definitions/model.Multimedia'
      multimedia_id:
        type: integer
    type: object
...
  model.CreateGoalRequest:
    properties:
      description:
        type: string
      end_date:
        type: string
      start_date:
        type: string
      target:
        minimum: 0
        type: integer
      title:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/model.GoalType'
        enum:
        - hafalan
        - khatam
        - tilawah
        - hadith
        - custom
    required:
    - start_date
    - title
    - type
    type: object
...
  model.CreateIslamicTermRequest:
    properties:
      category:
        $ref: '#/definitions/model.TermCategory'
      definition:
        type: string
      example:
        type: string
      origin:
        type: string
      source:
        type: string
      term:
        type: string
    required:
    - definition
    - term
    type: object
...
  model.GoalType:
    enum:
    - hafalan
    - khatam
    - tilawah
    - hadith
    - custom
    type: string
    x-enum-varnames:
    - GoalTypeHafalan
    - GoalTypeKhatam
    - GoalTypeTilawah
    - GoalTypeHadith
    - GoalTypeCustom
  model.Hadith:
    properties:
      book:
        $ref: '#/definitions/model.Book'
      book_id:
        type: integer
      chapter:
        $ref: '#/definitions/model.Chapter'
      chapter_id:
        type: integer
      defaultLanguage:
        type: string
      dhaif_by:
        type: string
      grade:
        allOf:
        - $ref: '#/definitions/model.HadithGrade'
        description: Grading & authentication
      grade_notes:
        type: string
      id:
        type: integer
      media:
        items:
          $ref: '#/definitions/model.HadithAsset'
        type: array
      number:
        type: integer
      sanad:
        description: Sanad (chain of narrators)
        type: string
      shahih_by:
        type: string
      theme:
        $ref: '#/definitions/model.Theme'
      theme_id:
        type: integer
      translation:
        $ref: '#/definitions/model.Translation'
      translation_id:
        type: integer
    type: object
  model.HadithAsset:
    properties:
      hadith_id:
        type: integer
      id:
        type: integer
      multimedia:
        $ref: '#/definitions/model.Multimedia'
      multimedia_id:
        type: integer
    type: object
  model.HadithGrade:
    enum:
    - shahih
    - shahih_lighairihi
    - hasan
    - hasan_lighairihi
    - hasan_shahih
    - dhaif
    - dhaif_jiddan
    - munkar
    - maudhu
    - matruk
    - majhul
    type: string
    x-enum-varnames:
    - HadithGradeShahih
    - HadithGradeShahihLighairihi
    - HadithGradeHasan
    - HadithGradeHasanLighairihi
    - HadithGradeHasanShahih
    - HadithGradeDhaif
    - HadithGradeDhaifJiddan
    - HadithGradeMunkar
    - HadithGradeMaudhu
    - HadithGradeMatruk
    - HadithGradeMajhul
...
  model.Multimedia:
    properties:
      file_name:
        description: File Name
        type: string
      file_size:
        description: File Size
        format: float
        type: number
      format:
        description: Format
        type: string
      id:
        type: integer
      original_file_name:
        description: Original File Name
        type: string
      title:
        description: Title
        type: string
      translation:
        allOf:
        - $ref: '#/definitions/model.Translation'
        description: Translation
      translation_id:
        description: Translation ID
        type: integer
      url:
        description: Url
        example: protocol://domain.ltd/path/to/image.extension
        type: string
    type: object
...
  model.TermCategory:
    enum:
    - fiqh
    - aqidah
    - tasawuf
    - ulumul_quran
    - hadith
    - lainnya
    type: string
    x-enum-varnames:
    - TermCategoryFiqh
    - TermCategoryAqidah
    - TermCategoryTasawuf
    - TermCategoryUlumulQuran
    - TermCategoryHadith
    - TermCategoryLainnya
  model.Theme:
    properties:
      books:
        items:
          $ref: '#/definitions/model.Book'
        type: array
      chapters:
        items:
          $ref: '#/definitions/model.Chapter'
        type: array
      default_language:
        type: string
      hadiths:
        items:
          $ref: '#/definitions/model.Hadith'
        type: array
      id:
        type: integer
      media:
        items:
          $ref: '#/definitions/model.ThemeAsset'
        type: array
      total_hadith:
        type: integer
      translation:
        $ref: '#/definitions/model.Translation'
      translation_id:
        type: integer
    type: object
  model.ThemeAsset:
    properties:
      id:
        type: integer
      multimedia:
        $ref: '#/definitions/model.Multimedia'
      multimedia_id:
        type: integer
      theme_id:
        type: integer
    type: object
  model.Translation:
    properties:
      ar:
        type: string
      ar_format:
        type: string
      ar_html:
        type: string
      ar_waqaf:
        type: string
      description_ar:
        type: string
      description_en:
        type: string
      description_idn:
        type: string
      en:
        type: string
      id:
        type: integer
      idn:
        type: string
      latin_en:
        type: string
      latin_idn:
        type: string
    type: object
...
  model.UpdateGoalRequest:
    properties:
      description:
        type: string
      end_date:
        type: string
      is_completed:
        type: boolean
      progress:
        minimum: 0
        type: integer
      target:
        minimum: 0
        type: integer
      title:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/books:
    get:
      consumes:
      - application/json
      description: Get paginated list of all books
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
      summary: List all books
      tags:
      - Books
    post:
      consumes:
      - application/json
      description: Create a new book entry
      parameters:
      - description: Book data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Book'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Create book
      tags:
      - Books
  /api/v1/books/{id}:
    delete:
      consumes:
      - application/json
      description: Delete book by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Delete book
      tags:
      - Books
    get:
      consumes:
      - application/json
      description: Get a single book by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get book by ID
      tags:
      - Books
    put:
      consumes:
      - application/json
      description: Update book by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Book data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Book'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Update book
      tags:
      - Books
  /api/v1/books/slug/{slug}:
    get:
      consumes:
      - application/json
      description: Get a single book by its slug
      parameters:
      - description: Book slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get book by slug
      tags:
      - Books
...
  /api/v1/goals:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get all study goals
      tags:
      - Belajar
    post:
      consumes:
      - application/json
      parameters:
      - description: Goal data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateGoalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Create study goal
      tags:
      - Belajar
  /api/v1/goals/{id}:
    delete:
      parameters:
      - description: Goal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Delete study goal
      tags:
      - Belajar
    put:
      consumes:
      - application/json
      parameters:
      - description: Goal ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateGoalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Update study goal
      tags:
      - Belajar
...
  /dictionary:
    get:
      consumes:
      - application/json
      parameters:
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Search keyword
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get all dictionary terms
      tags:
      - Belajar
    post:
      consumes:
      - application/json
      parameters:
      - description: Term data
        in: body
        name: term
        required: true
        schema:
          $ref: '#/definitions/model.CreateIslamicTermRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Create a dictionary term
      tags:
      - Belajar
  /dictionary/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Term ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Delete a dictionary term
      tags:
      - Belajar
    put:
      consumes:
      - application/json
      parameters:
      - description: Term ID
        in: path
        name: id
        required: true
        type: integer
      - description: Term data
        in: body
        name: term
        required: true
        schema:
          $ref: '#/definitions/model.CreateIslamicTermRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Update a dictionary term
      tags:
      - Belajar
  /dictionary/{term}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Term name
        in: path
        name: term
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get dictionary term by term
      tags:
      - Belajar
  /dictionary/category/{category}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Category name
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get dictionary terms by category
      tags:
      - Belajar
...
swagger: "2.0"